documentos dentro de las carpetas sin poder descargar aun

+++ api/CEDigital.API/Models/ArchivoDto.cs
@@ -7,7 +7,7 @@ namespace CEDigital.API.Models
         public int IdArchivo { get; set; }
         public required string NombreArchivo { get; set; }
         public DateTime FechaPublicacion { get; set; }
-        public long TamañoArchivo { get; set; }
+        public int TamañoArchivo { get; set; }
         public int IdCarpeta { get; set; }
         public string? Descripcion { get; set; }
         public required string Ruta { get; set; } // O solo el nombre del archivo si la ruta es interna
